refactor(plugin-text): extract dynamic placeholder functionality
